add UUID validation

+++ controllers/concerts.controller.js
@@ -1,4 +1,5 @@
 const Concert = require('../models/concert.model');
+const mongoose = require('mongoose');
 
 exports.getAll = async (req, res) => {
   try {
@@ -10,15 +11,17 @@ exports.getAll = async (req, res) => {
 };
 
 exports.getById = async (req, res) => {
-
-  try {
-    const con = await Concert.findById(req.params.id);
-    if (!con) res.status(404).json({ message: 'Not found' });
-    else res.json(con);
-  }
-  catch (err) {
-    res.status(500).json({ message: err });
-  }
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(501).json({ message: 'Invalid UUID' });
+  } else
+    try {
+      const con = await Concert.findById(req.params.id);
+      if (!con) res.status(404).json({ message: 'Not found' });
+      else res.json(con);
+    }
+    catch (err) {
+      res.status(500).json({ message: err });
+    }
 
 };
 
@@ -39,38 +42,42 @@ exports.post = async (req, res) => {
 
 exports.put = async (req, res) => {
   const { performer, genre, price, day, image } = req.body;
-
-  try {
-    const con = await Concert.findById(req.params.id);
-    if (con) {
-      con.performer = performer;
-      con.genre = genre;
-      con.day = day;
-      con.price = price;
-      con.image = image;
-      await con.save();
-      res.json({ message: 'OK' });
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(501).json({ message: 'Invalid UUID' });
+  } else
+    try {
+      const con = await Concert.findById(req.params.id);
+      if (con) {
+        con.performer = performer;
+        con.genre = genre;
+        con.day = day;
+        con.price = price;
+        con.image = image;
+        await con.save();
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
-  }
-  catch (err) {
-    res.status(500).json({ message: err });
-  }
+    catch (err) {
+      res.status(500).json({ message: err });
+    }
 
 };
 
 exports.delete = async (req, res) => {
-
-  try {
-    const con = await Concert.findById(req.params.id);
-    if (con) {
-      await Concert.deleteOne({ _id: req.params.id });
-      res.json({ message: 'OK' });
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(501).json({ message: 'Invalid UUID' });
+  } else
+    try {
+      const con = await Concert.findById(req.params.id);
+      if (con) {
+        await Concert.deleteOne({ _id: req.params.id });
+        res.json({ message: 'OK' });
+      }
+      else res.status(404).json({ message: 'Not found...' });
     }
-    else res.status(404).json({ message: 'Not found...' });
-  }
-  catch (err) {
-    res.status(500).json({ message: err });
-  }
+    catch (err) {
+      res.status(500).json({ message: err });
+    }
 
 };
